CRUD with authentication and authorization

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModal')
+const User = require('../models/userModel')
 
 // @desc Get Goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler (async (req,res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user:req.user.id})
 
     res.status(200).json(goals)
 })
@@ -19,7 +20,8 @@ const setGoal = asyncHandler (async (req,res) => {
         throw new Error('Please add a text field')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user:req.user.id
     })
     res.status(200).json(goal)
 })
@@ -32,6 +34,19 @@ const updateGoal = asyncHandler (async (req,res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoals = await Goal.findByIdAndUpdate(req.params.id,req.body,{
@@ -50,7 +65,19 @@ const deleteGoal = asyncHandler (async (req,res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+    const user = await User.findById(req.user.id)
 
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
     await goal.deleteOne()
 
     res.status(200).json({ id:req.params.id })
